AtS E1S3: fixed the unintended visible floating crown in the bottom right corner of the game map.
This is done by using a Lua hack that ensures the side 6 leader
placeholder unit stays hidden when starting the scenario or reloading
saved games (preload event handler). This is the [hide_unit] action
minus the implicit framebuffer flush (see its Lua implementation) --
this allows us to change the unit's hidden state without temporarily
locking the viewport at the top left corner of the map before display
initialization has finished.

It's not possible to just use [hide_unit] once and be done with it; it
sets .status.hidden = yes on affected units, but this status is
apparently ignored upon reloading (bug?).

[Wesnoth-UMC-Dev SVN r13010]