AtS: update the Aragwaith unit graphics so they match their MP faction add-on counterparts
This is the graphics portion of the change series. All graphics have
been run through the wesnoth-optipng script beforehand.

The MP faction [1] and patch were provided by bumbadadabum and use
wayfarer's latest iteration of the Aragwaith graphics [2].

 [1] http://r.wesnoth.org/t36901
 [2] http://r.wesnoth.org/p529256

This commit replaces the old graphics in AtS with the ones from
bumbadadabum's add-on entirely, except for the male Sorcerer (who is
still lingering around for some reason), and the campaign-specific Seer
unit (who is actually an early iteration of the female Wizard, also by
wayfarer).

[Wesnoth-UMC-Dev SVN r14801]