AtS E1: convert to use the new title cards format
In this commit, the title cards are placed so they are shown on the very
last storyscreen for each segment. E3 will be converted to follow this
convention later.

Note that there are no title cards for E1S12 or E1S13, and that E1S11.1
and E1S11.2 are treated as a single scenario. This is deliberate.

[Wesnoth-UMC-Dev SVN r16927]