AtS: use [unit] to spawn spawner spawns instead of relying upon undefined [unstore_unit] behavior
The previous approach involved generating partial unit state by hand in
a WML container and assuming that [unstore_unit] would do the right
thing afterwards. However, this does not seem to be always the case
(e.g. unusual behavior for impassable locations), so we might as well
take a safer approach for this task.

[Wesnoth-UMC-Dev SVN r16961]