AtS: remove redundant opacity-shifted versions of touchplate-100.png (now touchplate.png)
This is a historical relic of IftU's development around Wesnoth 1.3.8
and later. I implemented the opacity-shift image path function (~O())
for Wesnoth 1.5.7, which makes the multiple image files redundant when
the game engine can deal with this at runtime with negligible overhead
thanks to the image cache mechanism.

The ITEM_TOUCHPLATE macro in macros/items.cfg has been updated to use
the image path function for this instead of the separate image files.

[Wesnoth-UMC-Dev SVN r17298]