AtS: swap [combo_info_dialog] listbox and multi_page parent columns
This fixes the assertion failure previous noted on r17295:

> wesnoth-1.10: src/gui/widgets/grid.cpp:544: virtual void gui2::tgrid::place(const gui2::tpoint&, const gui2::tpoint&): Assertion `false' failed.
> [1]    19973 abort      wesnoth-1.10 -d

Normally, the listbox should be on the left and the multi_page parent
grid on the right. However, this causes Wesnoth to crash due to running
out of space for the listbox because the multi_page scroll_label
children *somehow* take precedence when updating the widgets' geometry
as the multi_page data is filled in by the preshow() callback.

This issue is averted by having the listbox on the rightmost column, so
it takes precedence over the multi_page children whenever geometry
changes. This is good, because the scroll_label widgets can then spawn
their own scrollbars as needed.

Perhaps someone else may be able to provide a proper solution to this
problem in the future.

[Wesnoth-UMC-Dev SVN r17338]