AtS: add linebreaks at the start of the descriptions of three specific abilities
This avoids triggering false positives with the "Descriptions should no
longer include the name as the first line" warning on 1.11.1 and later.

The code warns about the presence of colons on the first line of an
ability/weapon special description, not just about whether it contains
the ability/weapon special's title, hence these three abilities were
triggering the warning.

[Wesnoth-UMC-Dev SVN r17344]