AtS: the post-Divergence side-switching workaround is only needed on 1.11.1 and 1.11.1+svn
Technically, only 1.11.1+svn up to r56420 needs the workaround, but we
have no way to check the game engine's SVN revision (and that isn't
always available at build time).

This commit adds a NEED_DIVERGENCE_SIDE_PERSISTENCE_WORKAROUND symbol
that is defined for versions [1.11.1, 1.11.2) and signals the
preprocessor to actually define the workaround macros used in
post-Divergence scenarios.

[Wesnoth-UMC-Dev SVN r17368]