AtS: don't return from [item_choice_dialog] when double-clicking list items
The current usage of this dialog in AtS doesn't have any side-effects,
and being able to double click list items to dismiss it may create an
impression of it having an effect on something.

For possible usages with side-effects, I should figure out a way to set
the return_value_id for the list rows' toggle_panel widgets at some
point; but again, that isn't needed in this campaign at this time.

[Wesnoth-UMC-Dev SVN r17421]