AtS: fixed several instances (E1S7, etc.) of sighted events taking place prematurely on Wesnoth 1.11.x
This is because of a change in the sighted event semantics allowing it
to fire for sides that do not have shroud or fog enabled. The most
obvious symptom was the spider dialogue in E1S7 taking place as soon as
the player recruited or recalled a unit on the first turn.

[Wesnoth-UMC-Dev SVN r17449]