AtS: experimental fix for the long-standing gate clipping issue when adjacent to stone wall corners
The issue is non-trivial and involves a bunch of fancy terrain graphics
WML (due to the way wall corners work) that I'm too lazy to figure out
right now, so for now we can test this ugly BLIT+CROP based hack.

A couple of caveats:
 * It is ugly and uses BLIT solely because I'm too lazy to do the
   necessary maths to _not_ use BLIT right now
 * Gates on vertical (N-S) corridors overlap with the wall transitions
   a in subtle manner, the overlapping portion depending on the gate's
   orientation

The hack basically works by splitting each gate overlay in two, with
each half being drawn on a different layer depending on the gate's
orientation. The default layer is used for most of the wall transitions,
so layer 1 is used for preventing the corner clipping issue.

Layer 1 cannot be used for _both_ fragments (otherwise I'd have used the
OVERLAY_L macro instead of this mess) because gates will be drawn on top
of the wrong wall transitions as well as the inner (black) wall
background, resulting in an even worse visual mess.

[Wesnoth-UMC-Dev SVN r17555]