AtS: fix a severe gold exploit affecting scenarios using the NO_RECRUITS_OR_RECALLS macro (currently only E3S6)
An unspecified change in unit recruitment/recall behavior in Wesnoth
version 1.11.1 and later results in recruitment/recall gold not being
deducted if the target unit does not pass prerecruit/prerecall (i.e. is
removed from the gamemap). As a result, the original implementation of
this macro effectively allows the player to increase their gold supply
as it was designed around the 1.10.x and earlier behavior where gold is
unconditionally deducted every time.

[Wesnoth-UMC-Dev SVN r19137]