AtS: add a check to the [hidden_unit] implementation to avoid clobbering existing on-map units
Ported from IftU r19205.

The ramifications of this oversight are serious enough to warrant a
changelog entry, plus a quick bugfix release: Elyssa or Elynia (or a
Fire Guardian) may be clobbered and lost forever due to [hidden_unit]
being used to spawn the Fallen Faeries at the start of stage 2 of the
boss fight in E3S10.

[Wesnoth-UMC-Dev SVN r19212]