AtS E3S4.1: fix a game-breaking side-effect from r19212 resulting in an unwinnable scenario
The placeholder enemy leader for side 9 is a Matrix Flow System, all of
which are currently unhittable (CTH overridden to 0%) in this scenario.
Because [hidden_unit] doesn't clobber units anymore and the
boss-spawning event actually relied on this behavior, side 9 was left
with an unhittable MFS leader in addition to the boss. This rendered the
scenario effectively unwinnable when there are any other standing enemy
leaders left at the time the boss is slain.

[Wesnoth-UMC-Dev SVN r19335]