Reserve 10 ms for sleeping near at the end of [fade_out_music]
This delay gives silence.ogg a chance to start playing before the volume
is reset, thus avoiding the noise from the previous track heard for an
instant after switching tracks on my system, at least.

I still suspect there is something wrong with the way music changes in
WML come into effect. All that this delay should do is give the UI and
game rendering code a chance to run, not affect the sound code in any
relevant way whatsoever. Of course, there is still evidence of the
underlying bug in the form of the noise heard at the beginning of E3S12
after silence.ogg is replaced with actual music, just to name a
particular instance of misbehavior.

The delay chosen for this workaround is the same minimum delay requested
from the display object by the implementation of wesnoth.delay().