E1S12: Restrict crawler spawning EH further than TQ_ENERGY_SUPPLY_INTACT_CONDITION
We can only really respawn crawlers when the matrices exists because the
code in this EH determines respawn locations according to the on-map
locations of the matrices, which aren't known before they are first
revealed. While we could address this design limitation, it doesn't
really seem worth the effort from a gameplay perspective seeing as how
there's only one turn before the reveal takes place, and spawning
crawlers around the pillars while the player is dealing with Shaxthal
creatures seems unfair and random.

This fixes the following warning:

  warning engine: variable_info: invalid WML array index, respawn_locs[

... apearing during the AI side 7 turn before the matrices are first
spawned.