E2S2: Fire Allyna's introduction when there are 2 or less bandits left
It turns out the AI may decide to retreat units when it considers the
situation to be hopeless. As a consequence, Allyna's introduction event
may not take place until much later once the player has managed to hunt
down the last bandit of the quintet.

Starting with 5 bandits, the AI will presumably only retreat when there
are less than 3 left, so this should be good enough for now. Hopefully
at some point I'll come up with a better algorithm to decide when Allyna
can be considered to be out of danger.