New hive spawner controller logic implemented in Lua
The new logic is contained within two WML tags: [spawner_spawn] and
[run_spawn_controller]. Both are intended exclusively for use as
implementation details of the existing SPAWNER_BASE and
SPAWN_CONTROLLER WML macros.

The goal of the new design is to decrease the amount of WML
(especially macro substitutions) required for every Shaxthal hive
secnario. The Lua code can be compiled and run far faster every time a
save is loaded and there are multiple steps in the implementation of
spawners that can be done more naturally in Lua (not all of them,
though).