E3S8C, E3S9, E3S10: Increase AI aggression from 1.0 to 9.0
According to mattsc, at aggression > 1.0, the AI will favor attacks
which increase chances of its own units being damaged by the opponent.
At this point of the campaign this should be a Good Thing™ since it
increases the chances the AI will lower Elynia or Elyssa's HP during
boss encounters while making it easier for the player to kill its units.