E3S9: Warp to Engineer Fendyrn's spawn first on 1.11.0+
Scroll immediately to the target location if we have the warp-scroll
feature (1.11.0 and later). We don't want to try to do that before
restoring the screen state on 1.10.x because scrolling could take any
amount of time depending on the user's preferences, and that could ruin
the intended WHITE_SCREEN flash effect.