Replace version checks with a single HAVE_SAVE_GOLD_ASPECT_BY_DEFAULT macro
This is used to signal when the Wesnoth version in use has the AI
recruitment_save_gold aspect turned on by default.

At this time, the versions for which we define this symbol are those in
the [1.11.7, 1.11.9) set. I believe someone said something about
re-evaluating the recruitment_save_gold impact on 1.13.x, so this macro
should stay regardless of any version compatibility changes affecting
1.11.x.