Fix side 1 dropping during certain scenario transitions in 1.11.13 and later
Adds a DO_NOT_REMOVE_FROM_CARRYOVER_ON_DEFEAT macro wrapping the
remove_from_carryover_on_defeat option (and its predecessor) for Wesnoth
1.11.13 and later, where it's needed to prevent the game from dropping
side 1's state (recall list and gold) at the end of a scenario when it's
considered 'defeated' because there are no leaders on the map.

In most of the affected scenarios, there is a side 1 leader in its
recall list, but the game doesn't consider that to be a valid excuse to
not consider side 1 to be defeated anymore.

For the record, this behavior change happened mid-feature freeze and
it's a deviation from 1.10.x behavior as well.