Work around unidentified E1S6.2 SIGSEGV bug caused by a small silence.ogg
... by providing a larger silence.ogg to take precedence over mainline's
using the binary_path mechanism. It seems that at least on my system,
SDL_mixer/libogg/libvorbis have a problem with repeated loading of very
small tracks (in file size, not length, probably).

This larger silence.ogg (40 KiB instead of mainline's 5 KiB) is
inaudible pink noise. If my conjecture is right, it serves as a crash
buffer so as to avoid accessing bad memory in library-land.

Note that silence.ogg is used by [fade_out_music] and the
ENDLEVEL_QUIET/ENDLEVEL_CONTINUE macros, hence it's relevance for the
crash scenario.

Nobody else has been able to reproduce this bug yet.