E2S10: Redesigned map past the central gates, reduced turn limit
This makes the second half of the scenario less tedious to complete, and
also easier. In particular, side 1 units enjoy greater mobility and the
ability to scout further ahead to prepare better for confronting the
side 6 leader, who is now *much* closer to the gates.

Various object and unit placements have changed as a result, also
removing about 40% of the drone spawns.