E2S8: Fix dead-woman-walking issues with Elynia during Ivyel's flight sequence
Elynia might be reduced to less than 1 HP before the event in question
is thrown from an 'attack end' handler. This results in her being
deleted from the map by the time the victory event is fired from here,
which is a Bad Thing™.

The "fix" is forcing her HP to be 1 or more in the middle of this
sequence. It's takes up less space than checking whether she's supposed
to die and killing her by hand, and completing this objective is already
hard enough as it is, so I might as well be charitable for once.

Reported by RatArmy on the forums: <http://r.wesnoth.org/p586871>