Remove facing randomization from FAKE_RECRUIT macro
At some point, [unit] was deliberately made to randomize unit facing
if left unspecified. This is the case in Wesnoth 1.11.x and later, so we
don't really need to do it by hand.

(Yes, vultraz, I know you are reading this and wondering if the macro is
used anywhere. It isn't. This may change later. Unused #defines don't
hurt anyone.)