Remove misleading and unnecessary side= attribute from boss AI WML
According to the wiki, [modify_ai] side= is not required when the tag is
used in the context of a side's initial [ai] definition. The AI engine
in question isn't always used for side=2 either (see E3S9, for example),
so we might as well avoid confusion and drop the attribute entirely.

<http://wiki.wesnoth.org/AiWML#Adding_and_Deleting_Aspects_with_the_.5Bmodify_ai.5D_Tag>