Preliminary attempt at animating Elynia's E1S6.1 cutscene
This uses the E1/E2 ranged attack animation set, which is actually quite
inadequate because of the fact that the accompanying narration states
she holds her fire in _both_ hands, and she cannot do that without
dropping the staff (see also ##shadowm conversation around this commit's
timestamp).

I'm committing this for now while I try to animate the scene as
originally intended or decide to give up and go with a single-hand
animation.