Add dummy campaign entry to use for the units tree generator
This makes it so the units tree generator sees a single campaign instead
of three separate ones. The dummy campaign defines symbols of all three
episodes, but Elynia's Sylvan Warden variation used in E3 is excluded
via an #ifdef check so it doesn't clobber the usual one from the
previous two episodes -- mostly because it constitutes a mild spoiler.