E3S9: Rewrite the Engineer boss backstabbing logic for Wesnoth 1.14
It appears that in Wesnoth 1.14, unit facings are set before the
'attack' event is dispatched as part of a fix for a bug where unit
facings wouldn't be changed at all if combat animations were disabled in
advanced preferences (pointed out by gfgtdf). We relied on this
semi-broken behaviour in Wesnoth 1.12, and as a result, we were
constantly checking whether Elynia was standing *at Elyssa's location*
on every attack strike with the Claw.

In order to keep track of where the Engineer is facing after every
interaction he can possibly have with Elynia we need to track moveto and
attack events for the both of them, increasing the event count by 4.

There's also some future-proofing added in this commit to make sure the
variable holding the Engineer's previous facing before being struck by
Elyssa always exists and that we don't try to use it in an SLF before
it's first set in combat/movement, just in case the devs decide to make
the SLF direction syntax stricter and we end up with new surprises later
down the line, e.g. in Wesnoth 1.16.

Closes #67.

This reminds me that the Engineer's can be very ambiguous in some
situations, e.g. s looks exactly the same as se. It might be a good idea
to add some kind of visual indication as to when it's Elyssa's time to
strike, at least once the player receives the hint from her about the
backstabbing mechanic.