Rewrite Stun weapon special implementation
This replaces most of the code with the 1.14 Dead Water version of the
weapon special, with the caveat that we retain the original exception
which renders the weapon special unusable on defense.

Now we use [object] duration=turn end instead of the convoluted
hand-editing mess we had before, but also very importantly, the weapon
special's effect is cleared off at the end of the turn of the affected
unit's side rather than at the start. This may have some balance
implications but also makes more sense than the previous behaviour.