Add the basis of the new character AMLA upgrades system
AMLA_VITALITY is a version of the default mainline AMLA that increases
the unit's XP pool by the same amount as other upgrades, used when there
are no other options available.

AMLA_TREE_LOCK is used to display a UI message indicating that the full
AMLA tree for the unit is not available yet (which is gonna be the case
for Elynia and Anya) and that more AMLAs will become available later.
"Later" in these campaigns means in later episodes. Elynia's AMLA tree
is divided in three parts, for example, and Anya's in two (for obvious
reasons).

AMLA_UI_HANDLER includes code for a global event that is supposed to
handle the outcome of AMLA_TREE_LOCK by undoing it and removing the
AMLA, and returning the player to the Advance Unit dialog. It's a bit
more complicated than necessary due to a bug in how
wesnoth.remove_modifications affects unrelated unit stats in 1.14.9 (and
presumably earlier versions as well).