Wire in Elynia's new character upgrades tree
This also decreases Elynia's initial XP pool from 150 to 64.

Her AMLA tree as of this writing is as follows:

  * Strength:
    * Strength I: HP +4, melee dmg +1
    * Strength II: HP +4, melee str +1
    * Strength III: HP +5
    * Strength IV (E2): HP +5, melee dmg +1
    * Strength V (E2): HP +7
  * Focus:
    * Focus I: ensnare str +1
    * Focus II: ensnare dmg +1 str +1
    * Focus III (E2): ensnare dmg +1
    * Focus IV (E2): ensnare dmg +1, stun weapon special
  * Shielding (requires Focus I):
    * Shielding I: arcane res +10%
    * Shielding II: impact res +10%
    * Shielding III (E2): cold res +10%, sylvan essence regen +1
    * Shielding IV (E3): fire res +10%
    * Shielding V (E3): protection ability
  * Thorns:
    * Thorns I: thorns dmg +2
    * Thorns II: thorns dmg +2, drains weapon special
  * Faerie Fire (requires Focus II, Strength II):
    * Faerie Fire I: replaces mystic fire, ranged/arcane 9-3 magical
    * Faerie Fire II (E2): faerie fire dmg +1
    * Faerie Fire III (E2): faerie fire str +1
    * Faerie Fire IV (E3) (req. Focus III, Shielding III, Strength III):
      faerie fire dmg +1
    * Faerie Fire V (E3) (req. Focus IV, Shielding IV, Strength IV):
      faerie fire dmg +1

NOTES:

This commit does not include the necessary changes to E2S7 to
force-enable her faerie fire I AMLA (if not obtained yet), nor the code
that will take care of making Elynia's AMLAs persist from Episode I to
Episode II. I haven't reviewed the Episode III AMLA setup code either to
make sure it will work correctly with this system.

An ad hoc anti-AMLA for part of the effects of Strength II may be
necessary for the beginning of Episode III so her melee attack doesn't
become absurd.

Some rebalancing of the XP increase for some or all AMLAs/AMLA tiers may
be necessary. I haven't decided this yet, but it's an important point to
address before wiring in Anya's AMLAs as well.

Finally, regardless of the AMLA situation I need to revisit the prose
for E1S6.1 since right now Elynia's faerie fire is a bit weaker than the
Sun Sylph's even when maxed out. Although that may be primarily because
the dextrous trait isn't taken into account (bug? see also the notes in
the implementation of AMLA_UI_HANDLER).