Add abscond ability
It allows the unit to survive one fatal hit every other turn, being left
at 8 HP and 2 MP less than the maximum so it can retreat to safety.

In reality the ability is going to be used for Anya's E3 AMLA tree and
it's tied to her nature as a Shapeshifter child.