Remove spurious duplicate Fire Wisp unit type
This unit type is skipped by the engine since the one defined in Naia
takes precedence. The only concrete differences between both involve
resistances and the L1 unit. For the sake of putting an end to this
long-standing divergence between AtS and IftU it's best that we use the
Naia version from now on.

It does however raise the question of why Elyssa normally summons Fire
Guardians, but in E3S7A.2 she can additionally summon Wisps that level
up to Fire Spirits instead (which have different resistances as well).
I may need to reconsider a few points about her summons as well as the
Naia Fire Wisp line itself.