E1S6.1: Fix flickering at the start of Elynia's animation
However.

Because of timing changes to WML actions in version 1.14 (which is the
whole reason for this commit as well), it's possible to briefly see her
standing animation right before the cutscene animation is triggered.

Unfortunately I can't think of an easy solution to this other than using
the recruit and death animations (kind of like we do with Ergea in
E3S5). That's gonna be a bit cumbersome but it seems like the only way
forward unless there's a secret mechanism to lock display updates
temporarily in Lua.