Carry over character XP ratios and [object]s
[object]s need to be carried over now since it is possible for Durvan to
obtain permanent objects in Episode 2. [apply_objects] in Naia takes
care of this for us, although it goes out of its way to skip objects
that don't have a description or aren't duration=forever, since there's
always a chance I might miss something and end up sending a unit with
temporary image mod effects or something into the cache (plus it's a
legitimate concern when certain weapon specials are involved).

The problem now, is that we need to calculate the unit's previous XP
ratio while we are rebuilding it. There's a good reason for this: the
rebuild process raises the XP pool ignoring any existing caps applied to
the unit. We cannot even re-apply the previous cap without messing
things up even more because the order in which it was applied to the
unit matters and we are not preserving the order of [modifications] in
[apply_amlas] and [apply_objects] (shame on me). So now
UPDATE_UNIT_FROM_GLOBAL_STORE has special knowledge about AMLA_XP_CAP
and records the previous XP ratio for it to peruse when calculating the
XP cap modification.

This is very awkwardly over-engineered overall and I'm sure at some
point nemaara is going to see this and yell at me for not using flat XP
increases in the first place.