E3S4.1: Fix demolition charge events
This allows declaring the expectation for the primary location not to be
a valid unit that will be treated as the source of the demolition
deaths. By doing this, we can omit the [secondary_unit] selection in the
main [kill] action, which will not match a unit and instead will cause a
WML Warning in version 1.14 (another one of those "convenient" changes
in semantics since version 1.12).

While we're at it, fix a heretofore unnoticed bug where the demolition
code wouldn't trigger events for the unit at the center of the explosion
when used in this mode. I guess nobody ever accidentally placed one of
the main characters on top of a primed charge in E3S4.1 before ending
their turn, huh?

Fixes #56.