E3S3: Rework Nar-hamoth miniboss (issue #53)
Along with the balancing changes to his unit type, this makes it
impossible to milk him for XP without betting on RNG throws. There isn't
a forced CTH protection against him anymore and you have to fight him
for 4 rounds (with some minor buffs to him on a couple of them) while
avoiding losses. Perhaps a little bit frustrating but it is supposed to
be a challenge.