Further refinements to Demolition's AoE terraform logic
There were some issues with commit
1bc59ccc83577c4cab8bd3f08fdfbffe9a87d08e causing unexpected behaviour in
E3S4 which uses the Demolition ability implementation against impassable
terrains. Fixed by making the exceptions against impassable and
unwalkable terrains apply only when the ability is being used by a unit.
This is a bit of an ugly hack but it works so heck.

Also made it so the default exceptions prevent water and swamp terrains
from being turned into dirt.