E2: Limit Zynara's control of Ghosts, disable Bats
This makes it so Zynara cannot recruit Vampire Bats, and she's also not
able to control more than a difficulty-dependent number of Ghost-type
units at a time. Attempting to recruit additional Ghosts causes the
player to lose control of a random Ghost from the map or recall list,
and disables recruitment of Ghosts (but not recalling) for the remainder
of the scenario.