E3S4.1: Fix siege charge inconsistencies
Firstly, fixed an issue with the scenario dating back to 2012 (!) that
caused charge_immune_locs to never be recorded because of an invalid
SLF.

Secondly, reworked the implementation of the demolition ability so the
extra SLF is *added* on top of the built-in SLFs, always in exclusion
mode. This means that it's no longer possible to invalidate the built-in
rules for the explosion radius.

Thirdly, the extra SLF is now used so locations in charge_immune_locs
are immune to the explosion effect.

Finally, we need to manually terraform the castle tiles underneath the
Matrix parts anyway because we don't want to leave those behind (they
ARE within the explosion radius), especially not with the impassable
overlay still applied. It's intentionally still possible for the player
to arrange charges so that some passable castle tiles remain for use
during the boss fight later on.

This whole ordeal was an exercise in frustration. How about not touching
this code ever again?