Don't bind a hotkey for Fire Guardian summoning
The action of binding a hotkey persists during the same game session,
and in this particular case since we're reusing a hotkey that's already
bound to an engine command, some dumb behaviour ensues when switching to
a scenario where the menu item doesn't exist and the player needs to use
the engine command again.

This bug was 7 years old.