E2S10: Aesthetic and balancing changes to the map and scenario
This makes it so the player has substantially less villages to work
with, reduced baseline income, and no initial gold bonus on top of
E2S9's carryover, making it so saving gold in E2S9 becomes a must.

See also issue #80.

(Some of the code affecting shroud removal during the initial cutscene
had to be changed to accomodate the aesthetic changes to the map.)