E2S11: Multiple boss fight changes for balancing purposes
This adds a stream of spawns from the abyss that includes both Shaxthal
and Demon units (including the occasional Demoness Spelldancer) in order
to increase the pressure on the player. Especially noteworthy is the
fact that this should render the Spectre meta less of a silver bullet
solution for this scenario.

I may still revisit some minor aspects in a later release.

Closes #80.