E3S11: Re-implement credits sequence fully in Lua
This moves almost all of the implementation to Lua and eliminates some
of the awkward sequencing syntax, replacing it with a single action WML
tag with more expressive screen structuring.

The WML containing the credits sequence is now in /about-final.cfg to
make it less cumbersome to edit as well (big indentation, etc.). This
file is in turn included by E3S11 only.