E1S9.2: Re-do terrain substitution mask for the Elyssa confrontation
This makes it so the terrain mask no longer has to be offset from the
map's origin point (1,1) and so border=yes is used consistently on 1.14.
It also brings a few visual improvements making the terrains used more
consistent with the campaign's current map aesthetic.

Elyssa's positioning has been changed to use her side's starting
location as well instead of the nonsensical hardcoded location we were
using (which to be fair, probably was intended to keep her closer to the
player's units vision range, but we don't really need that to be honest,
she's perfectly visible as it is).

Additionally, we spawn random debris with [item] in the same sequence.
Why [item] instead of the actual debris terrain? Because we already use
the terrain overlay layer to place Fake Fog initially, and then to
simulate the fog fade-out using Event Flag overlays in combination with
scenario-local [terrain_graphics] rules.

This should *hopefully* make the scenario 1.16-proof, but there's no way
to be sure without testing it later. There's already red flags in that
[terrain_mask]'s semantics were changed yet again, making border= a
no-op and requiring an alignment= attribute which we do not currently
specify (!!!). This is according to wiki.w.o/TerrainMaskWML as of this
writing, anyway.

(So yeah, this is probably broken in 1.16 right now. Crap.)