E3S11: Try to center credits text better
To this end, we try to guess the height of the overlay to compensate
for the fact that the engine API does not do that by itself. This is
just an educated guess and may not properly reflect the height of the
text, but it's better than no offset at the moment, since otherwise
text can easily spill off the screen at the bottom half.